Initial commit: Add three fragments with bottom navigation and ViewPager2
This commit adds the initial version of a learning project that uses fragments and Retrofit API. The project includes three fragments with a bottom navigation menu and ViewPager2. The app fetches data from the Dog API to display random images of dogs as a way to test the Retrofit API implementation.

The goal of this project is to learn how to use fragments and Retrofit API, as well as to get familiar with Android development best practices.

Changes in this commit:
- Add three fragments for displaying dog images
- Implement bottom navigation menu
- Implement ViewPager2 for swiping between fragments
- Add Retrofit API to fetch data from the Dog API
- Add placeholder images to be displayed while fetching data